Import: Modified such that authorize() is run on a separate thread, allowing us to set a timeout for the user to enter his account details Import: Extracted out methods relating to Google authentication/information retrieval into separate class GoogleUtil
